Move shader-param.ts + test into kv-core
